🔨 Pull login function out into a module
